removed HCA specifc code
