+++ Dockerfile
@@ -5,8 +5,7 @@ RUN mkdir /app
 WORKDIR /app
 
 ADD src ./src
-ADD config ./config
 COPY package*.json ./
 
 RUN npm install
-CMD ["node", "./src/app.js"]
+CMD ["node", "./src/server.js"]
