--- config/custom-environment-variables.json
@@ -1,41 +1,0 @@
-{
-  "AMQP" : {
-    "metadataValidation" : {
-      "connection" : {
-        "host" : "RABBIT_HOST",
-        "port" : "RABBIT_PORT"
-      }
-    },
-    "fileValidationResults" : {
-      "connection" : {
-        "host" : "RABBIT_HOST",
-        "port" : "RABBIT_PORT"
-      }
-    }
-  },
-  "INGEST_API" : {
-    "connection" : {
-      "host" : "INGEST_HOST",
-      "port" : "INGEST_PORT"
-    }
-  },
-  "UPLOAD_API" : {
-    "connection" : {
-      "scheme": "UPLOAD_SCHEME",
-      "host" : "UPLOAD_HOST",
-      "port" : "UPLOAD_PORT"
-    },
-    "apiKey" : "UPLOAD_API_KEY"
-  },
-  "OLS_API" : {
-    "connection" : {
-      "host" : "OLS_HOST",
-      "port" : "OLS_PORT"
-    }
-  },
-  "FILE_VALIDATION_IMAGES" : {
-    "fastq" : "FASTQ_VALIDATION_IMAGE",
-    "fastq.gz" : "FASTQ_VALIDATION_IMAGE",
-    "fastq.tar.gz" : "FASTQ_VALIDATION_IMAGE"
-  }
-}
