--- config/default.json
@@ -1,57 +1,0 @@
-{
-  "AMQP" : {
-    "metadataValidation" : {
-      "connection" : {
-        "scheme": "amqp",
-        "host" : "localhost",
-        "port" : 5672
-      },
-      "messaging" : {
-        "exchange" : "ingest.validation.exchange",
-        "queueName" : "ingest.metadata.validation.queue",
-        "routingKey" : "ingest.metadata.validation.queue",
-        "exchangeType": "direct"
-      }
-    },
-    "fileValidationResults" : {
-      "connection" : {
-        "scheme": "amqp",
-        "host" : "localhost",
-        "port" : 5672
-      },
-      "messaging" : {
-        "exchange" : "ingest.validation.exchange",
-        "queueName" : "ingest.file.validation.queue",
-        "routingKey" : "ingest.file.validation.queue",
-        "exchangeType": "direct"
-      }
-    }
-  },
-  "INGEST_API" : {
-    "connection" : {
-      "scheme": "http",
-      "host" : "localhost",
-      "port" : 8080
-    }
-  },
-  "UPLOAD_API" : {
-    "connection" : {
-      "scheme": "https",
-      "host" : "upload.dev.data.humancellatlas.org",
-      "port" : 80
-    },
-    "apiKey" : "redacted"
-  },
-  "OLS_API" : {
-    "connection" : {
-      "scheme": "http",
-      "host" : "ontology.dev.data.humancellatlas.org",
-      "port" : 80
-    }
-  },
-  "FILE_VALIDATION_IMAGES" : {
-    "fastq" : "quay.io/humancellatlas/fastq_utils:master",
-    "fastq.gz" : "quay.io/humancellatlas/fastq_utils:master",
-    "fastq.tar.gz" : "quay.io/humancellatlas/fastq_utils:master"
-  }
-}
