+++ package.json
@@ -1,31 +1,31 @@
 {
-  "name": "json-schema-validator",
-  "version": "1.5.2",
-  "description": "A nodejs JSON schema validator.",
-  "main": "src/server.js",
-  "repository": "https://github.com/EMBL-EBI-SUBS/json-schema-validator.git",
+  "name": "elixir-ajv-keywords",
+  "version": "1.0.0",
+  "description": "Custom JSON-Schema keywords for Ajv validator",
+  "main": "index.js",
+  "repository": "https://github.com/elixir-europe/elixir-ajv-keywords.git",
   "scripts": {
     "start": "node src/server.js",
     "test": "jest",
     "test-watch": "nodemon --exec \"npm test\"",
-    "travis-test": "jest test/validator.test.js test/schema-references.test.js"
+    "travis-test": "jest test/elixir-validator.test.js"
   },
   "engines": {
     "node": "8.11.1"
   },
   "keywords": [
     "nodejs",
+    "biojs",
     "json",
     "schema",
     "validator"
   ],
-  "author": "EMBL-EBI-SUBS, fpenim",
+  "author": "EMBL-EBI-SUBS, fpenim, rolando-ebi, simonjupp",
   "license": "Apache-2.0",
   "dependencies": {
     "ajv": "^6.0.0",
     "amqplib": "^0.5.2",
     "bluebird": "^3.5.1",
-    "config": "^2.0.1",
     "dnscache": "^1.0.1",
     "express": "^4.16.2",
     "express-validator": "^5.2.0",
@@ -42,5 +42,9 @@
   "devDependencies": {
     "jest": "^22.4.3",
     "nodemon": "^1.17.5"
+  },
+  "jest": {
+    "verbose": true,
+    "testURL": "http://localhost/"
   }
 }
