+++ src/keywords/isvalidterm.js
