--- src/utils/ingest-client/ingest-client-exceptions.js
@@ -1,9 +1,0 @@
-/**
- * Created by rolando on 06/08/2018.
- */
-
-class NoUuidError extends Error {}
-
-module.exports = {
-    "NoUuidError" : NoUuidError
-};
