--- src/validation/validation-handler.js
@@ -1,5 +1,0 @@
-const validate = require("./validator-prototype");
-
-exports.handleValidation = function (schemas, entity, rootSchemaId = schemas[0].$id) {
-    return validate(schemas, entity, rootSchemaId);
-}
