+++ test/async-validation.test.js
@@ -1,94 +1,110 @@
 const fs = require("fs");
-const runValidation = require("../src/validation/validator");
+const ElixirValidator = require('../src/elixir-validator');
+const IsChildTerm = require('../src/keywords/ischildtermof');
 
 test(" -> isChildTermOf Schema", () => {
-  let inputSchema = fs.readFileSync("examples/schemas/ischildterm-schema.json");
-  let jsonSchema = JSON.parse(inputSchema);
+    let inputSchema = fs.readFileSync("examples/schemas/isChildTerm-schema.json");
+    let jsonSchema = JSON.parse(inputSchema);
 
-  let inputObj = fs.readFileSync("examples/objects/isChildTerm.json");
-  let jsonObj = JSON.parse(inputObj);
+    let inputObj = fs.readFileSync("examples/objects/isChildTerm.json");
+    let jsonObj = JSON.parse(inputObj);
 
-  return runValidation.validateSingleSchema(jsonSchema, jsonObj).then( (data) => {
-    expect(data).toBeDefined();
-    expect(data[0]).toBeDefined();
-    expect(data[0].dataPath).toBe(".attributes['age'][0].terms[0].url");
-  });
+    const elixirValidator = new ElixirValidator([IsChildTerm]);
+
+    return elixirValidator.validate(jsonSchema, jsonObj).then( (data) => {
+        expect(data).toBeDefined();
+        expect(data.validationErrors.length).toBe(1);
+        expect(data.validationErrors[0].absoluteDataPath).toBe(".attributes['age'][0].terms[0].url");
+    });
 });
 
 test("FAANG Schema - FAANG \'organism\' sample", () => {
-  let inputSchema = fs.readFileSync("examples/schemas/faang-schema.json");
-  let jsonSchema = JSON.parse(inputSchema);
+    let inputSchema = fs.readFileSync("examples/schemas/faang-schema.json");
+    let jsonSchema = JSON.parse(inputSchema);
 
-  let inputObj = fs.readFileSync("examples/objects/faang-organism-sample.json");
-  let jsonObj = JSON.parse(inputObj);
+    let inputObj = fs.readFileSync("examples/objects/faang-organism-sample.json");
+    let jsonObj = JSON.parse(inputObj);
 
-  return runValidation.validateSingleSchema(jsonSchema, jsonObj).then( (data) => {
-    expect(data).toBeDefined();
-    expect(data.length).toBe(0);
-  });
+    const elixirValidator = new ElixirValidator([IsChildTerm]);
+
+    return elixirValidator.validate(jsonSchema, jsonObj).then( (data) => {
+
+        expect(data).toBeDefined();
+        expect(data.validationErrors.length).toBe(0);
+    });
 });
 
 test("FAANG Schema - \'specimen\' sample", () => {
-  let inputSchema = fs.readFileSync("examples/schemas/faang-schema.json");
-  let jsonSchema = JSON.parse(inputSchema);
+    let inputSchema = fs.readFileSync("examples/schemas/faang-schema.json");
+    let jsonSchema = JSON.parse(inputSchema);
 
-  let inputObj = fs.readFileSync("examples/objects/faang-specimen-sample.json");
-  let jsonObj = JSON.parse(inputObj);
+    let inputObj = fs.readFileSync("examples/objects/faang-specimen-sample.json");
+    let jsonObj = JSON.parse(inputObj);
 
-  return runValidation.validateSingleSchema(jsonSchema, jsonObj).then( (data) => {
-    expect(data).toBeDefined();
-    expect(data.length).toBe(0);
-  });
+    const elixirValidator = new ElixirValidator([IsChildTerm]);
+
+    return elixirValidator.validate(jsonSchema, jsonObj).then( (data) => {
+        expect(data).toBeDefined();
+        expect(data.validationErrors.length).toBe(0);
+    });
 });
 
 test("FAANG Schema - \'pool of specimens\' sample", () => {
-  let inputSchema = fs.readFileSync("examples/schemas/faang-schema.json");
-  let jsonSchema = JSON.parse(inputSchema);
+    let inputSchema = fs.readFileSync("examples/schemas/faang-schema.json");
+    let jsonSchema = JSON.parse(inputSchema);
 
-  let inputObj = fs.readFileSync("examples/objects/faang-poolOfSpecimens-sample.json");
-  let jsonObj = JSON.parse(inputObj);
+    let inputObj = fs.readFileSync("examples/objects/faang-poolOfSpecimens-sample.json");
+    let jsonObj = JSON.parse(inputObj);
 
-  return runValidation.validateSingleSchema(jsonSchema, jsonObj).then( (data) => {
-    expect(data).toBeDefined();
-    expect(data.length).toBe(0);
-  });
+    const elixirValidator = new ElixirValidator([IsChildTerm]);
+
+    return elixirValidator.validate(jsonSchema, jsonObj).then( (data) => {
+        expect(data).toBeDefined();
+        expect(data.validationErrors.length).toBe(0);
+    });
 });
 
 test("FAANG Schema - \'cell specimen\' sample", () => {
-  let inputSchema = fs.readFileSync("examples/schemas/faang-schema.json");
-  let jsonSchema = JSON.parse(inputSchema);
+    let inputSchema = fs.readFileSync("examples/schemas/faang-schema.json");
+    let jsonSchema = JSON.parse(inputSchema);
 
-  let inputObj = fs.readFileSync("examples/objects/faang-cellSpecimen-sample.json");
-  let jsonObj = JSON.parse(inputObj);
+    let inputObj = fs.readFileSync("examples/objects/faang-cellSpecimen-sample.json");
+    let jsonObj = JSON.parse(inputObj);
 
-  return runValidation.validateSingleSchema(jsonSchema, jsonObj).then( (data) => {
-    expect(data).toBeDefined();
-    expect(data.length).toBe(0);
-  });
+    const elixirValidator = new ElixirValidator([IsChildTerm]);
+
+    return elixirValidator.validate(jsonSchema, jsonObj).then( (data) => {
+        expect(data).toBeDefined();
+        expect(data.validationErrors.length).toBe(0);
+    });
 });
 
 test("FAANG Schema - \'cell culture\' sample", () => {
-  let inputSchema = fs.readFileSync("examples/schemas/faang-schema.json");
-  let jsonSchema = JSON.parse(inputSchema);
+    let inputSchema = fs.readFileSync("examples/schemas/faang-schema.json");
+    let jsonSchema = JSON.parse(inputSchema);
 
-  let inputObj = fs.readFileSync("examples/objects/faang-cellCulture-sample.json");
-  let jsonObj = JSON.parse(inputObj);
+    let inputObj = fs.readFileSync("examples/objects/faang-cellCulture-sample.json");
+    let jsonObj = JSON.parse(inputObj);
 
-  return runValidation.validateSingleSchema(jsonSchema, jsonObj).then( (data) => {
-    expect(data).toBeDefined();
-    expect(data.length).toBe(0);
-  });
+    const elixirValidator = new ElixirValidator([IsChildTerm]);
+
+    return elixirValidator.validate(jsonSchema, jsonObj).then( (data) => {
+        expect(data).toBeDefined();
+        expect(data.validationErrors.length).toBe(0);
+    });
 });
 
 test("FAANG Schema - \'cell line\' sample", () => {
-  let inputSchema = fs.readFileSync("examples/schemas/faang-schema.json");
-  let jsonSchema = JSON.parse(inputSchema);
+    let inputSchema = fs.readFileSync("examples/schemas/faang-schema.json");
+    let jsonSchema = JSON.parse(inputSchema);
 
-  let inputObj = fs.readFileSync("examples/objects/faang-cellLine-sample.json");
-  let jsonObj = JSON.parse(inputObj);
+    let inputObj = fs.readFileSync("examples/objects/faang-cellLine-sample.json");
+    let jsonObj = JSON.parse(inputObj);
 
-  return runValidation.validateSingleSchema(jsonSchema, jsonObj).then( (data) => {
-    expect(data).toBeDefined();
-    expect(data.length).toBe(0);
-  });
+    const elixirValidator = new ElixirValidator([IsChildTerm]);
+
+    return elixirValidator.validate(jsonSchema, jsonObj).then( (data) => {
+        expect(data).toBeDefined();
+        expect(data.validationErrors.length).toBe(0);
+    });
 });
