+++ test/graphRestriction.test.js
@@ -1,5 +1,7 @@
 const fs = require("fs");
-const runValidation = require("../src/validation/validator");
+const ElixirValidator = require('../src/elixir-validator');
+const GraphRestriction = require('../src/keywords/graph_restriction');
+
 
 test(" -> graphRestriction Schema", () => {
     let inputSchema = fs.readFileSync("examples/schemas/graphRestriction-schema.json");
@@ -8,9 +10,12 @@ test(" -> graphRestriction Schema", () => {
     let inputObj = fs.readFileSync("examples/objects/graphRestriction_pass.json");
     let jsonObj = JSON.parse(inputObj);
 
-    return runValidation.validateSingleSchema(jsonSchema, jsonObj).then( (data) => {
+    const ingestValidator = new ElixirValidator([GraphRestriction]);
+
+    return ingestValidator.validate(jsonSchema, jsonObj).then((data) => {
         expect(data).toBeDefined();
     });
+
 });
 
 test(" -> graphRestriction Schema", () => {
@@ -20,7 +25,9 @@ test(" -> graphRestriction Schema", () => {
     let inputObj = fs.readFileSync("examples/objects/graphRestriction_normal.json");
     let jsonObj = JSON.parse(inputObj);
 
-    return runValidation.validateSingleSchema(jsonSchema, jsonObj).then( (data) => {
+    const ingestValidator = new ElixirValidator([GraphRestriction]);
+
+    return ingestValidator.validate(jsonSchema, jsonObj).then((data) => {
         expect(data).toBeDefined();
     });
 });
@@ -32,8 +39,12 @@ test(" -> graphRestriction Schema", () => {
     let inputObj = fs.readFileSync("examples/objects/graphRestriction_fail.json");
     let jsonObj = JSON.parse(inputObj);
 
-    return runValidation.validateSingleSchema(jsonSchema, jsonObj).then( (data) => {
+    const ingestValidator = new ElixirValidator([GraphRestriction]);
+
+    return ingestValidator.validate(jsonSchema, jsonObj).then((data) => {
         expect(data).toBeDefined();
-        expect(data[0]).toBeDefined();
+        expect(data.validationErrors.length).toBe(1);
+        expect(data.validationErrors[0].userFriendlyMessage).toContain('Provided term is not child of');
+
     });
 });
